Modified - move wastes provider
