+++ lib/features/waste_feature/presentation/wastes_screen.dart
@@ -6,7 +6,7 @@ import 'package:tamizshahr/features/waste_feature/business/entities/wasteCart.da
 
 
 import '../../../core/theme/app_theme.dart';
-import '../../../provider/wastes.dart';
+import 'providers/wastes.dart';
 import '../../../core/widgets/main_drawer.dart';
 import 'widgets/waste_item_wastes_screen.dart';
