+++ lib/features/waste_feature/presentation/widgets/waste_cart_item.dart
@@ -7,7 +7,7 @@ import 'package:tamizshahr/features/waste_feature/business/entities/wasteCart.da
 
 
 import '../../../../core/theme/app_theme.dart';
-import '../../../../provider/wastes.dart';
+import '../providers/wastes.dart';
 import '../../../../core/logic/en_to_ar_number_convertor.dart';
 
 class WasteCartItem extends StatefulWidget {
