+++ lib/features/waste_feature/presentation/widgets/waste_cart_item_animated_list.dart
@@ -6,7 +6,7 @@ import 'package:tamizshahr/features/waste_feature/business/entities/price_weight
 import 'package:tamizshahr/features/waste_feature/business/entities/wasteCart.dart';
 
 import '../../../../core/theme/app_theme.dart';
-import '../../../../provider/wastes.dart';
+import '../providers/wastes.dart';
 import '../../../../core/logic/en_to_ar_number_convertor.dart';
 
 class WasteCartItemAnimatedList extends StatefulWidget {
