+++ lib/main.dart
@@ -16,7 +16,7 @@ import 'package:tamizshahr/features/waste_feature/presentation/wastes_screen_ani
 import './provider/articles.dart';
 import 'features/customer_feature/presentation/providers/auth.dart';
 import './provider/messages.dart';
-import './provider/wastes.dart';
+import 'features/waste_feature/presentation/providers/wastes.dart';
 import './screens/about_us_screen.dart';
 import 'features/waste_feature/presentation/address_screen.dart';
 import './screens/article_detail_screen.dart';
