+++ lib/screens/collect_detail_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:intl/intl.dart' as intl;
 import 'package:provider/provider.dart';
 import 'package:tamizshahr/features/waste_feature/business/entities/request_waste_item.dart';
-import 'package:tamizshahr/provider/wastes.dart';
+import 'package:tamizshahr/features/waste_feature/presentation/providers/wastes.dart';
 import 'package:tamizshahr/widgets/collect_details_collects_item.dart';
 import 'package:tamizshahr/core/logic/en_to_ar_number_convertor.dart';
